+++ src/components/3d/Scene.tsx
@@ -4,11 +4,14 @@ import Flights from "./Flights.tsx";
 import FlightTrail from "./FlightTrail.tsx";
 import {Suspense} from "react";
 import {EARTH_RADIUS} from "../../constants.ts";
-import {Bloom, EffectComposer, Vignette} from "@react-three/postprocessing";
+import {Bloom, EffectComposer, SMAA, Vignette} from "@react-three/postprocessing";
+import {useRecoilValue} from "recoil";
+import {graphicOptionsState} from "../../atoms.ts";
 import {MobileEarth} from "./MobileEarth.tsx";
 
 
 function Scene() {
+    const graphicOptions = useRecoilValue(graphicOptionsState)
 
     return (
         <Suspense fallback={
@@ -35,10 +38,27 @@ function Scene() {
             />
             <Flights/>
             <FlightTrail/>
-            <EffectComposer>
-                <Bloom luminanceThreshold={0} luminanceSmoothing={0.9} height={300}/>
-                <Vignette eskil={false} offset={0.1} darkness={0.9}/>
-            </EffectComposer>
+            {
+                graphicOptions.bloom || graphicOptions.vignette || graphicOptions.SMAA ?
+                    <EffectComposer>
+                        {
+                            graphicOptions.bloom ?
+                                <Bloom luminanceThreshold={0} luminanceSmoothing={0.9} height={300}
+                                /> : <></>
+                        }
+                        {
+                            graphicOptions.vignette ?
+                                <Vignette eskil={false} offset={0.1} darkness={0.9}
+                                /> : <></>
+                        }
+                        {
+                            graphicOptions.SMAA ?
+                                <SMAA/> : <></>
+                        }
+
+                    </EffectComposer>
+                    : <></>
+            }
         </Suspense>
     )
 }